test(grid): extract shared grid perf test helpers into utils
