+++ tests/grid-perf/grid-perf-app.spec.ts
@@ -1,4 +1,10 @@
-import { test, type Page } from '@playwright/test';
+import { test } from '@playwright/test';
+import {
+  logResults,
+  scrollGrid,
+  startFPSMeasurement,
+  stopFPSMeasurement,
+} from './utils';
 
 /**
  * Grid Performance Tests using the standalone perf app.
@@ -19,117 +25,6 @@ import { test, type Page } from '@playwright/test';
 
 const PERF_APP_URL = 'http://localhost:4020';
 
-interface FPSResult {
-  fps: number;
-  avgFrameTime: number;
-  minFrameTime: number;
-  maxFrameTime: number;
-  frameCount: number;
-  droppedFrames: number;
-}
-
-async function startFPSMeasurement(page: Page): Promise<void> {
-  await page.evaluate(() => {
-    (window as any).__frameTimings = [];
-    (window as any).__fpsRunning = true;
-    let lastTime = performance.now();
-
-    function measureFrame() {
-      if (!(window as any).__fpsRunning) return;
-
-      const now = performance.now();
-      (window as any).__frameTimings.push(now - lastTime);
-      lastTime = now;
-      requestAnimationFrame(measureFrame);
-    }
-    requestAnimationFrame(measureFrame);
-  });
-}
-
-async function stopFPSMeasurement(page: Page): Promise<FPSResult> {
-  const timings = await page.evaluate(() => {
-    (window as any).__fpsRunning = false;
-    return (window as any).__frameTimings as number[];
-  });
-
-  const validTimings = timings.filter(t => t > 0);
-
-  if (validTimings.length === 0) {
-    return {
-      fps: 0,
-      avgFrameTime: 0,
-      minFrameTime: 0,
-      maxFrameTime: 0,
-      frameCount: 0,
-      droppedFrames: 0,
-    };
-  }
-
-  const avgFrameTime =
-    validTimings.reduce((a, b) => a + b, 0) / validTimings.length;
-  const fps = 1000 / avgFrameTime;
-  const minFrameTime = Math.min(...validTimings);
-  const maxFrameTime = Math.max(...validTimings);
-  const droppedFrames = validTimings.filter(t => t > 33).length;
-
-  return {
-    fps,
-    avgFrameTime,
-    minFrameTime,
-    maxFrameTime,
-    frameCount: validTimings.length,
-    droppedFrames,
-  };
-}
-
-/**
- * Scrolls the grid in the perf app using mouse wheel events
- */
-async function scrollPerfAppGrid(
-  page: Page,
-  totalDelta: number
-): Promise<void> {
-  const canvas = page.locator('canvas').first();
-  const box = await canvas.boundingBox();
-  if (!box) throw new Error('Grid canvas not found');
-
-  await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
-
-  const scrollStep = 100;
-  const direction = Math.sign(totalDelta);
-  let remaining = Math.abs(totalDelta);
-
-  while (remaining > 0) {
-    const step = Math.min(scrollStep, remaining);
-    await page.mouse.wheel(0, step * direction);
-    remaining -= step;
-    await page.waitForTimeout(16);
-  }
-}
-
-function logResults(
-  testName: string,
-  result: FPSResult,
-  expected: { minFps: number }
-): void {
-  console.log(`\n${testName}:`);
-  console.log(`  Average FPS: ${result.fps.toFixed(1)}`);
-  console.log(`  Avg frame time: ${result.avgFrameTime.toFixed(2)}ms`);
-  console.log(
-    `  Frame time range: ${result.minFrameTime.toFixed(
-      2
-    )}ms - ${result.maxFrameTime.toFixed(2)}ms`
-  );
-  console.log(`  Total frames: ${result.frameCount}`);
-  console.log(
-    `  Dropped frames (>33ms): ${result.droppedFrames} (${(
-      (result.droppedFrames / result.frameCount) *
-      100
-    ).toFixed(1)}%)`
-  );
-  console.log(`  Expected min FPS: ${expected.minFps}`);
-}
-
 test.describe('grid perf app - stress tests', () => {
   test.skip(
     !process.env.RUN_PERF_TESTS,
@@ -144,10 +39,11 @@ test.describe('grid perf app - stress tests', () => {
 
     await startFPSMeasurement(page);
 
-    await scrollPerfAppGrid(page, 5000);
-    await scrollPerfAppGrid(page, -3000);
-    await scrollPerfAppGrid(page, 4000);
-    await scrollPerfAppGrid(page, -5000);
+    const canvas = page.locator('canvas').first();
+    await scrollGrid(page, canvas, 5000);
+    await scrollGrid(page, canvas, -3000);
+    await scrollGrid(page, canvas, 4000);
+    await scrollGrid(page, canvas, -5000);
 
     const result = await stopFPSMeasurement(page);
     logResults('1M Rows Scroll', result, { minFps: 30 });
